Move errorsync back to internal
