+++ internal/errorsync/err.go
